+++ src/components/controls/slider.vue
@@ -1,10 +1,10 @@
 <template>
-  <div class="control-slider-container" @mouseleave="dragleave" @mousemove="drag" @mouseup="dragend" @mouseenter="dragend">
+  <div class="control-slider-container" @mousemove="drag" @mouseleave="dragleave" @mouseup="dragend" @mouseenter="dragend">
     <div class="slide-bar" ref="slider">
       <div class="fader" v-bind:style="offsetstyle" @mousedown.prevent="dragstart" @mouseup="dragend" ref="fader"></div>
     </div>
     <div class="text-container" :data-preunit="preunit" :data-postunit="postunit">
-      <control-text class="small" v-model="m" :disabled="true" />
+      <control-text class="small" v-model="display" :disabled="true" />
     </div>
   </div>
 </template>
@@ -47,11 +47,12 @@ export default {
   },
   data() {
     return {
-      m: `${this.preunit}${this.preunit ? ' ' : ''}${this.processDisplay(this.value > this.range[0] ? this.value : this.range[0])} ${this.postunit}`,
+      display: null,
       dragrestart: true,
       dragstate: false,
       dragorigin: 0,
       dragoffset: 0,
+      nextOffset: 0,
       dragposition: 0,
       dragstop: true,
       offsetstyle: {left: 0},
@@ -66,9 +67,6 @@ export default {
     dragstart(e) {
       this.dragstate = true;
       this.dragorigin = e.screenX;
-      this.sliderBarWidth = this.$refs.slider.clientWidth;
-      this.faderWidth = this.$refs.fader.clientWidth;
-      this.maxDrag = this.sliderBarWidth - this.faderWidth - this.boundryPad;
     },
     dragend(e) {
       this.dragstate = false;
@@ -79,28 +77,48 @@ export default {
     },
     drag(e) {
       if (this.dragstate) {
-        let nextOffset =  e.screenX - this.dragorigin + this.dragposition;
-        if (nextOffset < 0 || nextOffset > this.maxDrag) {
+        this.nextOffset =  e.screenX - this.dragorigin + this.dragposition;
+        let tickVal = scale(0, this.maxDrag, this.range[0], this.range[1], 1);
+
+        // if (nextVal < this.range[0] || nextVal > this.range[1]) {
+        if (this.nextOffset <= 0 || this.nextOffset >= this.maxDrag) {
           this.dragstop = true;
         } else {
           this.dragstop = false;
-          this.dragoffset = nextOffset;
-          let val = scale(0, this.maxDrag, this.range[0], this.range[1], this.dragoffset)
-          let displaynum = this.processDisplay(val);
-          this.m = `${this.preunit}${this.preunit ? ' ' : ''}${displaynum} ${this.postunit}`;
-          this.offsetstyle = {left: `${this.dragoffset}px`};
-          this.updateValue(val)
+          this.dragoffset = this.nextOffset;
+          this.offsetstyle = {left: `${Math.floor(this.dragoffset)}px`};
+          let val = scale(0, this.maxDrag, this.range[0], this.range[1], this.dragoffset);
+          if (val >= this.range[0]-tickVal && val <= this.range[1]+tickVal) {
+            this.updateValue(val);
+          }
         }
       }
     },
     updateValue: function(val) {
-      this.val = val;
-      this.$emit('input', val);
+      let displaynum = this.processDisplay(val);
+      this.display = `${this.preunit}${this.preunit ? ' ' : ''}${displaynum} ${this.postunit}`;
+      this.val = parseFloat(val.toFixed(this.places + Math.log10(this.factor)));
+      this.$emit('input', this.val);
     },
     processDisplay(val) {
       let adjustedVal = val * this.factor;
       return fixedNotation(adjustedVal, this.places, config.get('utils.denominations'));
+    },
+    initSlider(e) {
+      this.display = `${this.preunit}${this.preunit ? ' ' : ''}${this.processDisplay(this.value)}${this.postunit ? ' ' : ''}${this.postunit}`,
+      this.sliderBarWidth = this.$refs.slider.clientWidth;
+      this.faderWidth = this.$refs.fader.clientWidth;
+      this.maxDrag = this.sliderBarWidth - this.faderWidth - this.boundryPad;
+      this.dragoffset = scale(...this.range, 0, this.maxDrag, this.value);
+      this.offsetstyle = {left: `${Math.floor(this.dragoffset)}px`};
     }
+  },
+  created() {
+    this.$parent.$on('animationcomplete', this.initSlider);
+    this.$parent.$on('reset', this.initSlider);
+  },
+  mounted() {
+    this.initSlider();
   }
 }
 </script>
@@ -112,7 +130,7 @@ export default {
     padding: 10px 0;
     @include flex-row-stack();
     .slide-bar {
-      flex: 8;
+      flex: 4;
       box-sizing: border-box;
       @include control-slider-bar($height: 1px, $background: $app-white, $border-width: 1px);
       .fader {
@@ -129,7 +147,7 @@ export default {
     .text-container {
       display: flex;
       align-items: center;
-      flex: 3;
+      flex: 1;
     }
   }
 </style>